+++ CHANGELOG.rst
@@ -17,6 +17,8 @@ Added
 - Add ``apply_first`` method to ``Apply``.
 - Add ``lift4`` and ``lift5`` functions.
 - Add ``flap`` (and ``flip``) functions.
+- Add ``UncurriedFunction``, ``UncurriedFunctionMonoid`` and ``uncurried`` in
+  order to support functions with arbitrary function signatures.
 
 Changed
 -------